Changed the velocity of ball as velocity(Vec2) into dir(Vec2) and float speed. Also added simple  rebounding factor according to the placement of the ball.

+++ Engine/Walls.cpp
@@ -14,10 +14,10 @@ Walls::Walls(Rect rect, float width)
 
 void Walls::Draw(Graphics & gfx) const
 {
-	gfx.DrawRect(rect.left - width, rect.top - width,	rect.right + width, rect.top ,				c);
-	gfx.DrawRect(rect.left - width, rect.top,			rect.left,			rect.bottom,			c);
-	gfx.DrawRect(rect.left - width, rect.bottom,		rect.right + width, rect.bottom + width ,	c);
-	gfx.DrawRect(rect.right,		rect.top,			rect.right + width, rect.bottom ,			c);
+	gfx.DrawRect(static_cast<int>(rect.left - width), static_cast<int>(rect.top - width),	static_cast<int>(rect.right + width),	static_cast<int>(rect.top) ,				c);
+	gfx.DrawRect(static_cast<int>(rect.left - width), static_cast<int>(rect.top),			static_cast<int>(rect.left),			static_cast<int>(rect.bottom),			c);
+	gfx.DrawRect(static_cast<int>(rect.left - width), static_cast<int>(rect.bottom),		static_cast<int>(rect.right + width),	static_cast<int>(rect.bottom + width) ,	c);
+	gfx.DrawRect(static_cast<int>(rect.right),			static_cast<int>(rect.top),			static_cast<int>(rect.right + width),	static_cast<int>(rect.bottom) ,			c);
 }
 
 Rect Walls::GetRect() const
